Refactor fzf package a little bit
